fix: usecallback problem solved
the problem: evertytime the screen Consultation was onFocus, the functin wrapped by usecallback was executed. This functin took the consultation global value and updated the local state. But this function neve updated, so the consultation global value awayls have the same value. The solution was: add at the usecallback function one depency, that was consultation global value. In this way, everytime consultiton global value changes, the usecallback function does the same, and the consutation global value in it changes too